requirements.txt
asgiref==3.7.2
Django==4.2.7
django-restframework==0.0.1
Pillow==10.1.0
pybind11==2.11.1
sqlparse==0.4.4
